Plume-backend-main/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
clickclick==20.10.2
ConfigArgParse==1.7
connexion==2.14.2
cryptography>=3.2
dataclasses-json==0.5.13
distro==1.8.0
ecdsa==0.18.0
exceptiongroup==1.2.0
Flask==2.2.5
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
frozenlist==1.4.0
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.0.239
langsmith==0.0.14
locust==2.20.0
Markdown==3.5.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==1.3.5
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.21.12
psutil==5.9.7
psycogreen==1.0.2
pyasn1==0.5.0
pydantic==1.10.11
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyMySQL==1.1.0
python-jose==3.3.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
roundrobin==0.0.4
rpds-py==0.9.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
stripe==7.6.0
swagger-ui-bundle==0.0.9
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
waitress==2.1.2
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.1
